backend/requirements.txt
django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
celery==5.3.6
redis==5.0.7
django-celery-results==2.5.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.131.0
groq==0.9.0
python-dotenv==1.0.1
cryptography==42.0.8
pydantic==2.7.1
django-environ==0.11.2
psycopg2-binary==2.9.9
